Engraver: Improve algorithm so stage moves only to neccessary points
Explanation of the algorithm:

init_steps() first calculates a coordinate matrix from the image
content setting the element value to 1 if the pixel brightness is 255
else 0.
Afterwards a distance list is calculated form this matrix which
contains the relative distances from one matrix element with the value
1 to the next one (those where the stage will move to).

In the next_move() loop calls the distance list is looped over and the
final distances considering the spot size are calculated.